Move dao test to dao test folder
