+++ app/src/androidTest/kotlin/com/jehutyno/yomikata/dao/KanjiSoloDaoTest.kt
@@ -1,10 +1,10 @@
-package com.jehutyno.yomikata.repository.local
+package com.jehutyno.yomikata.dao
 
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
-import com.jehutyno.yomikata.dao.KanjiSoloDao
+import com.jehutyno.yomikata.repository.local.*
 import org.junit.Assert.*
 
 import org.junit.After
