+++ app/src/androidTest/kotlin/com/jehutyno/yomikata/dao/WordDaoTest.kt
@@ -1,11 +1,12 @@
-package com.jehutyno.yomikata.repository.local
+package com.jehutyno.yomikata.dao
 
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
-import com.jehutyno.yomikata.dao.QuizDao
-import com.jehutyno.yomikata.dao.WordDao
+import com.jehutyno.yomikata.repository.local.RoomQuizWord
+import com.jehutyno.yomikata.repository.local.RoomWords
+import com.jehutyno.yomikata.repository.local.YomikataDataBase
 import org.junit.Assert.*
 
 import org.junit.After
